requirement.txt
bcrypt==3.1.7
Bootstrap-Flask==1.3.2
boto3==1.14.23
botocore==1.17.23
Brotli==1.0.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dash==1.12.0
dash-bootstrap-components==0.10.5
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
dnspython==1.16.0
docutils==0.15.2
email-validator==1.1.1
Flask==1.1.2
Flask-Compress==1.5.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
kiwisolver==1.2.0
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.0
newsapi-python==0.2.6
numpy==1.18.5
pandas==1.0.4
patsy==0.5.1
Pillow==7.2.0
plotly==4.8.1
progressbar2==3.51.4
psycopg2-binary==2.8.5
pycparser==2.20
pygam==0.8.0
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
requests==2.24.0
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.5.1
seaborn==0.10.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.17
statsmodels==0.11.1
threadpoolctl==2.1.0
ua-parser==0.10.0
urllib3==1.25.9
user-agents==2.2.0
Werkzeug==1.0.1
WTForms==2.3.1
yapf==0.30.0